Downloads/AI-VOICE-AGENT/requirements.txt
sounddevice
soundfile
numpy
openai-whisper
pyttsx3
llama-cpp-python
torch
tqdm
faiss-cpu
sentence-transformers
transformers
